+++ packages/sveltekit/src/client/index.ts
@@ -26,18 +26,38 @@ const generateId = (): string => {
 	return generateRandomNumber().toString(36).slice(2, 7);
 };
 
+export class UndefinedError extends Error {
+	constructor(
+		type: "pageData._lucia" | "_luciaStore" | "_setLuciaStore" | "_luciaPageData" | "_lucia"
+	) {
+		const errorMsg = {
+			"pageData._lucia":
+				"page data property _lucia is undefined  - Make sure handleServerSession(auth) is set up inside the root +layout.server.ts",
+			_luciaStore:
+				"global variable _luciaStore is undefined - Make sure handleHooks(auth) is set up inside hooks.server.ts",
+			_setLuciaStore:
+				"global variable _setLuciaStore is undefined - Make sure handleHooks(auth) is set up inside hooks.server.ts",
+			_luciaPageData:
+				"global variable _luciaPageData is undefined - Make sure handleHooks(auth) is set up inside hooks.server.ts",
+			_lucia:
+				"_lucia context does not exist in page data - Make sure handleSession() set inside the root +layout.svelte file"
+		} as const;
+		super(errorMsg[type]);
+	}
+}
+
 export const handleSession = (
 	pageStore: Readable<{
 		data: PageData;
 	}>,
 	onSessionUpdate: (hasSession: boolean) => void = () => {}
 ) => {
-	if (typeof window === "undefined") return;
+	if (typeof document === "undefined") return;
 	const broadcastChannel = new BroadcastChannel("__lucia__");
 	const tabId = generateId();
 	const initialPageStoreValue = get(pageStore);
 	const initialLuciaContext = initialPageStoreValue.data._lucia;
-	if (!initialLuciaContext) throw new Error("pageData._lucia is undefined");
+	if (!initialLuciaContext) throw new UndefinedError("pageData._lucia");
 	const globalWindow = window as GlobalWindow;
 	let pageStoreUnsubscribe = () => {},
 		userStoreUnsubscribe = () => {};
@@ -50,7 +70,7 @@ export const handleSession = (
 	globalWindow._luciaStore = readable<LuciaContext>(initialLuciaContext, (set) => {
 		globalWindow._setLuciaStore = set;
 	});
-	if (!globalWindow._luciaStore) throw new Error("_luciaStore is undefined");
+	if (!globalWindow._luciaStore) throw new UndefinedError("_luciaStore");
 	userStoreUnsubscribe = globalWindow._luciaStore.subscribe((newContext) => {
 		/*
 		prevent postMessage on store initialization
@@ -63,8 +83,8 @@ export const handleSession = (
 	});
 	pageStoreUnsubscribe = pageStore.subscribe((pageStoreValue) => {
 		const newLuciaContext = pageStoreValue.data?._lucia;
-		if (!newLuciaContext) throw new Error("pageData._lucia is undefined");
-		if (!globalWindow._setLuciaStore) throw new Error("_setLuciaStore() is undefined");
+		if (!newLuciaContext) throw new UndefinedError("pageData._lucia");
+		if (!globalWindow._setLuciaStore) throw new UndefinedError("_setLuciaStore");
 		globalWindow._setLuciaStore(newLuciaContext);
 	});
 	broadcastChannel.addEventListener("message", ({ data }) => {
@@ -76,7 +96,7 @@ export const handleSession = (
 		check if message is coming from the same tab
 		*/
 		if (messageData.tabId === tabId) return;
-		if (!globalWindow._luciaStore) throw new Error("_luciaStore is undefined");
+		if (!globalWindow._luciaStore) throw new UndefinedError("_luciaStore");
 		const currentLuciaContext = get(globalWindow._luciaStore);
 		/*
 		check if session has changed from the previous page data