+++ packages/lucia-auth/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.1
 
 - Remove node dependencies (`crypto`, `util`)
+- Adds `@noble/hashes` as dependency
 - Use block size (`r`) of `16` for hashing passwords
 - Add `configs.hash.generate()` and `configs.hash.validate()` for custom hashing implementation
 - Normalize password string on hashing