+++ packages/lucia-auth/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.2.1
 
-- Remove node dependencies (`crypto`, `util`)
+- Remove node dependencies (`crypto`, `util`) [#236](https://github.com/pilcrowOnPaper/lucia-auth/issues/236)
 - Adds `@noble/hashes` as dependency
 - Use block size (`r`) of `16` for hashing passwords
 - Add `configs.hash.generate()` and `configs.hash.validate()` for custom hashing implementation